moving functions per comments
