+++ tests/automated/event-render/EventRenderUtils.js
@@ -1,7 +1,10 @@
-import { getEventEls } from '../lib/MonthViewUtils'
+const EVENT_CLASS = 'fc-event';
+const TITLE_CLASS = 'fc-title';
+const RESIZER_CLASS = 'fc-resizer';
+const TIME_CLASS = 'fc-time';
 
 export function expectIsStart(bool) {
-  var el = getSingleEl()
+  var el = getSingleEl();
 
   if (bool) {
     expect(el).toHaveClass('fc-start')
@@ -11,13 +14,46 @@ export function expectIsStart(bool) {
 }
 
 export function expectIsEnd(bool) {
-  var el = getSingleEl()
+  var el = getSingleEl();
 
   if (bool) {
     expect(el).toHaveClass('fc-end')
   } else {
     expect(el).not.toHaveClass('fc-end')
   }
+}
+
+export function getEventElTimeText(el) {
+  return $(el).find(`.${TIME_CLASS}`).text()
+}
+
+export function getVisibleEventEls() {
+  return $(`.${EVENT_CLASS}:visible`)
+}
+
+export function getEventEls(eventClassName = EVENT_CLASS) {
+  return $(`.${eventClassName}`)
+}
+
+export function getSingleEventEl(eventEl) {
+  expect(eventEl.length).toBe(1)
+  return eventEl
+}
+
+export function getFirstEventEl(eventClassName) {
+  return getSingleEventEl(getEventEls(eventClassName).first())
+}
+
+export function getLastEventEl(eventClassName) {
+  return getSingleEventEl(getEventEls(eventClassName).last())
+}
+
+export function getEventElTitleEl(eventEl) {
+  return eventEl.find(`.${TITLE_CLASS}`)
+}
+
+export function getEventElResizerEl(eventEl) {
+  return eventEl.find(`.${RESIZER_CLASS}`)
 }
 
 export function getSingleEl() {
