+++ tests/automated/event-render/TimeGridEventRenderUtils.js
@@ -1,5 +1,6 @@
-import {getDayEls, getEventEls, getEventElTimeText} from '../lib/MonthViewUtils';
+import {getDayEls} from '../view-render/DayGridRenderUtils';
 import {getTimeGridSlotEls} from '../lib/time-grid';
+import {getEventEls, getEventElTimeText} from "./EventRenderUtils";
 
 export function getTimeTexts() {
   return getEventEls().map(function(i, eventEl) {
